* Created a ResponseHelper to provide some utilities for HTTP responses like:     * error messages     * json parsing     * returning null objects * Applied some fixes:     * changed int to long for IDs

+++ src/main/java/com/github/glo2003/daos/IDAO.java
@@ -1,10 +1,10 @@
 package com.github.glo2003.daos;
 
 public interface IDAO<T> {
-  T get(int id);
+  T get(long id);
   // List<T> getAll();
 
-  int save(T t);
+  long save(T t);
   // void update(T t, T newT);
   // void delete(T t);
   // void delete(long id);
