+++ src/main/java/com/github/glo2003/helpers/MathHelper.java
@@ -3,7 +3,7 @@ package com.github.glo2003.helpers;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class MathHelper {
-  public static int randomInt() {
-    return ThreadLocalRandom.current().nextInt(0, Integer.MAX_VALUE - 1);
+  public static long randomLong() {
+    return ThreadLocalRandom.current().nextLong(0, Long.MAX_VALUE - 1);
   }
 }
